# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	components/chainhook-event-observer/Cargo.toml
#	components/chainhook-event-observer/doc/openapi.json
#	components/chainhook-event-observer/src/chainhooks/mod.rs
#	components/chainhook-event-observer/src/indexer/bitcoin/blocks_pool.rs
#	components/chainhook-event-observer/src/indexer/stacks/blocks_pool.rs
#	components/chainhook-event-observer/src/indexer/stacks/mod.rs
#	components/chainhook-event-observer/src/indexer/tests/helpers/bitcoin_blocks.rs
#	components/chainhook-event-observer/src/indexer/tests/helpers/microblocks.rs
#	components/chainhook-event-observer/src/indexer/tests/helpers/stacks_blocks.rs
#	components/chainhook-event-observer/src/indexer/tests/helpers/transactions.rs
#	components/chainhook-event-observer/src/observer/mod.rs
#	components/chainhook-event-observer/src/observer/tests/mod.rs
#	components/clarinet-cli/src/frontend/dap.rs
#	components/clarinet-cli/src/integrate/chains_coordinator.rs
#	components/clarinet-deployments/Cargo.toml
#	components/clarinet-deployments/src/lib.rs
#	components/clarinet-files/src/lib.rs
#	components/clarinet-files/src/network_manifest.rs
#	components/clarity-lsp/Cargo.toml
#	components/clarity-lsp/src/common/state.rs
#	components/clarity-repl/src/clarity/util/address/c32.rs
#	components/clarity-repl/src/clarity/util/c32.rs
#	components/clarity-repl/src/clarity/util/hash.rs
#	components/clarity-repl/src/codec/mod.rs
#	vendor/orchestra-event-observer/src/chainhooks/mod.rs
#	vendor/orchestra-event-observer/src/indexer/stacks/mod.rs

+++ components/clarinet-cli/src/frontend/telemetry.rs
@@ -11,7 +11,7 @@ pub enum DeveloperUsageEvent {
     NewProject(DeveloperUsageDigest),
     PokeExecuted(DeveloperUsageDigest),
     CheckExecuted(DeveloperUsageDigest),
-    TestSuiteExecuted(DeveloperUsageDigest, bool, u32),
+    TestSuiteExecuted(DeveloperUsageDigest, bool, usize),
     DevnetExecuted(DeveloperUsageDigest),
     ProtocolPublished(DeveloperUsageDigest, StacksNetwork),
     DebugStarted(DeveloperUsageDigest, u32),
